+++ src/core-taggeds-optional/Optional/Present.T/Present.T.Factory.Constructor.cs
@@ -1,0 +1,11 @@
+#nullable enable
+
+namespace System
+{
+    partial struct Present<T>
+    {
+        public Present(T value)
+            =>
+            this.value = value;
+    }
+}
