+++ src/core-taggeds-optional/Optional/Present.T/Present.T.Factory.Explicit.cs
@@ -1,0 +1,11 @@
+#nullable enable
+
+namespace System
+{
+    partial struct Present<T>
+    {
+        public static Present<T> From(T value)
+            =>
+            new(value);
+    }
+}
