+++ src/core-taggeds-optional/Optional/Present.T/Present.T.Factory.Implicit.cs
@@ -1,0 +1,11 @@
+#nullable enable
+
+namespace System
+{
+    partial struct Present<T>
+    {
+        public static implicit operator Present<T>(T value)
+            =>
+            new(value);
+    }
+}
