+++ src/core-taggeds-optional/Optional/Present.T/Present.T.Value.Explicit.cs
@@ -1,0 +1,9 @@
+#nullable enable
+
+namespace System
+{
+    partial struct Present<T>
+    {
+        public T Value => value;
+    }
+}
