+++ src/core-taggeds-optional/Optional/Present.T/Present.T.Value.Implicit.cs
@@ -1,0 +1,11 @@
+#nullable enable
+
+namespace System
+{
+    partial struct Present<T>
+    {
+        public static explicit operator T(Present<T> present)
+            =>
+            present.value;
+    }
+}
